Tests/LibWeb/Text/expected/css/font-face-load-dedups-same-url.txt

faceA.status: loaded
faceB.status: loaded
resource entries total: 3
resource entries for font url: 2
